Only load rake tasks if rake DSL is present
It's reasonably possible for `Rake` to be defined, but the DSL hasn't been 
included into the global namespace. If this is the case (such as happens 
when RubyMine runs tests within the IDE), then this line will blow up when
jettywrapper gets required. Here we add an extra test to make sure that the
required DSL is in place before we load the tasks.