Load rake tasks only if defined?(Rake::VERSION)
In jettywrapper.rb, only load rake tasks if defined?(Rake::VERSION).
Testing only defined?(Rake) is brittle because some code libraries
monkey patch Rake.  If the Rake gem is not loaded, this patching
will nevertheless cause the Rake constant to spring into existence,
and jettywrapper will mistakenly then attempt to load its Rake tasks,
even though the core Rake functionality is absent.

A real-world case of this is the JetBrains TeamCity code that is
automatically loaded when line debugging Ruby code in RubyMine/Intellij.
If the rake gem is not loaded for a given Run/Debug configuration,
the session will error out inside the jettywrapper gem.

Closes #53